# public/openapi.yaml
openapi: 3.0.3
info:
  title: Ottodot Trial Booking API
  version: "1.0.0"
  description: >
    Backend API for trial-class booking only (no regular enrollment).
    Authoritative behavior is documented in TECH_DESIGN.md; this spec describes
    the request/response contracts for every endpoint listed there (§7).
servers:
  - url: http://localhost:3000
    description: Local dev server
security:
  - sessionCookie: []
...
paths:
...
  /api/bookings:
    get:
      operationId: listMyBookings
      tags: [bookings]
      summary: List all of the logged-in parent's own bookings
      description: >
        Every booking (any status) across all of the caller's children —
        "which classes has my family already booked and paid for". Scoped
        entirely by session; there is no parentId parameter.
      responses:
        "200":
          description: The caller's bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MyBookingEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create (or reuse) a trial booking
      description: >
        Runs, in order: session/role check, ownership check (studentId belongs to the
        caller), isTrial eligibility check, duplicate-booking check, and the Step A
        optimistic capacity pre-check (TECH_DESIGN.md §4-§5). Returns the existing
        pending_payment booking idempotently if one already exists for this
        student+class pair.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created (or existing pending booking returned)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Already booked, or class is full
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                alreadyBooked:
                  value: { error: already_booked }
                classFull:
                  value: { error: class_full }
        "422":
          description: Class is not currently offered as a trial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notATrialClass:
                  value: { error: not_a_trial_class }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_token
...
  responses:
    Unauthorized:
      description: No valid session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unauthorized:
              value: { error: unauthorized }
    Forbidden:
      description: Authenticated, but not allowed to access this resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            forbidden:
              value: { error: forbidden }
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notFound:
              value: { error: not_found }
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - already_booked
            - class_full
            - not_a_trial_class
            - card_declined
            - capacity_exceeded
            - not_found
            - forbidden
            - unauthorized
            - email_taken
            - invalid_credentials
            - not_cancellable
...
    BookingStatus:
      type: string
      enum: [pending_payment, confirmed, payment_failed, cancelled]
    Booking:
      type: object
      properties:
        id:
          type: string
        studentId:
          type: string
        classId:
          type: string
        status:
          $ref: "#/components/schemas/BookingStatus"
        failureReason:
          type: string
          nullable: true
          enum: [capacity_exceeded, card_declined, expired_timeout, duplicate, null]
        cancellationReason:
          type: string
          nullable: true
          description: Free-text reason an admin gave when cancelling this booking (POST /api/bookings/:id/cancel). Distinct from failureReason, which is system-generated codes.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateBookingRequest:
      type: object
      required: [studentId, classId]
      properties:
        studentId:
          type: string
        classId:
          type: string
...
    MyBookingEntry:
      type: object
      properties:
        id:
          type: string
        status:
          $ref: "#/components/schemas/BookingStatus"
        failureReason:
          type: string
          nullable: true
        cancellationReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        student:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
        classData:
          type: object
          properties:
            id:
              type: string
            subject:
              type: string
            startsAt:
              type: string
              format: date-time